refactor: разложить классы по пакетам
