+++ src/main/java/ru/practicum/shareit/user/model/User.java
@@ -1,4 +1,4 @@
-package ru.practicum.shareit.user;
+package ru.practicum.shareit.user.model;
 
 /**
  * TODO Sprint add-controllers.
